backend/pyproject.toml
[project]
name = "director-hub-api"
version = "0.1.0"
description = "Director Hub PRO — FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.32.0,<1.0",
    "python-jose[cryptography]>=3.3.0,<4.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "httpx>=0.28.0,<1.0",
    "gpxpy>=1.6.0,<2.0",
    "numpy>=2.0,<3.0",
    "plotly>=6.0,<7.0",
    "python-multipart>=0.0.9,<1.0",
    "supabase>=2.0,<3.0",
    "python-dotenv>=1.0,<2.0",
    "reportlab>=4.0,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
